+++ src/FluxFlow.Components.State/Options/StateComponentOptions.cs
@@ -1,4 +1,5 @@
 using FluxFlow.Components.Expressions;
+using FluxFlow.Components.State.Timing;
 using FluxFlow.Engine.Mapping;
 
 namespace FluxFlow.Components.State.Options;
@@ -6,6 +7,9 @@ namespace FluxFlow.Components.State.Options;
 public sealed class StateComponentOptions
 {
     private readonly FlowExpressionEngineRegistry _expressionEngines = new("State");
+    private IStateClock _clock = SystemStateClock.Instance;
+
+    public IStateClock Clock => _clock;
 
     public StateComponentOptions UseExpressionEngine(
         IFlowExpressionEngine expressionEngine,
@@ -22,6 +26,12 @@ public sealed class StateComponentOptions
         Func<string?, IFlowExpressionEngine> resolver)
     {
         _expressionEngines.UseResolver(resolver);
+        return this;
+    }
+
+    public StateComponentOptions UseClock(IStateClock clock)
+    {
+        _clock = clock ?? throw new ArgumentNullException(nameof(clock));
         return this;
     }