+++ src/FluxFlow.Components.State/Timing/IStateClock.cs
@@ -1,0 +1,6 @@
+namespace FluxFlow.Components.State.Timing;
+
+public interface IStateClock
+{
+    DateTimeOffset UtcNow { get; }
+}